🐭 Shape can be moved and bbox updates
