+++ src/global/classes/canvasObjects/__test__/MarcherShape.test.ts
@@ -1,0 +1,3 @@
+import { describe } from "vitest";
+
+describe.todo("MarcherShape");
